add cutlass qlinear
