+++ .gitmodules
@@ -1,3 +1,3 @@
-[submodule "third_party/cudnn-frontend"]
-	path = third_party/cudnn-frontend
-	url = https://github.com/NVIDIA/cudnn-frontend.git
+[submodule "third_party/cutlass"]
+	path = third_party/cutlass
+	url = https://github.com/NVIDIA/cutlass.git
