+++ sfast/csrc/main.cpp
@@ -3,8 +3,8 @@
 #include "jit/init.h"
 #include "misc.h"
 #include "operators/cudnn/cudnn_convolution.h"
-#include "operators/cudnn/cudnn_qlinear.h"
 #include "operators/cublas/cublas_gemm.h"
+#include "operators/cutlass/cutlass_qlinear.h"
 #include "operators/fused_linear.h"
 
 namespace sfast {
@@ -16,8 +16,8 @@ PYBIND11_MODULE(TORCH_EXTENSION_NAME, m) {
 
 TORCH_LIBRARY(sfast, m) {
   operators::initCUDNNConvolutionBindings(m);
-  operators::initCUDNNQLinearBindings(m);
   operators::initCUBLASGEMMBindings(m);
+  operators::initCutlassQLinearBindings(m);
   operators::initFusedLinearBindings(m);
 }
