--- sfast/csrc/operators/cudnn/cudnn_qlinear.cc
@@ -1,693 +1,0 @@
-#include <torch/extension.h>
-
-// #include <ATen/native/cudnn/Macros.h>
-#include <c10/util/ArrayRef.h>
-
-#include <ATen/ATen.h>
-#include <ATen/cuda/Exceptions.h>
-#include <ATen/cudnn/Handle.h>
-#include <ATen/cudnn/Types.h>
-// #include <ATen/native/quantized/cudnn/utils.h>
-#include <ATen/TensorUtils.h>
-#include <ATen/native/quantized/PackedParams.h>
-#include <ATen/native/utils/ParamsHash.h>
-#include <c10/core/ScalarType.h>
-#include <c10/cuda/CUDACachingAllocator.h>
-#include <c10/cuda/CUDAFunctions.h>
-#include <cudnn_frontend.h>
-
-#include <iostream>
-#include <unordered_map>
-
-#include "cudnn_qlinear.h"
-#include "cudnn_quantized_utils.h"
-#include "cudnn_utils.h"
-
-namespace {
-// TODO: there is a table from input dtype and weight dtype to operator dtype,
-// we can derive the operator dtype based on input dtype
-cudnn_frontend::MatMulDesc_v8 getLinearDescriptor(cudnnDataType_t dataType) {
-  return cudnn_frontend::MatMulDescBuilder().setMathPrecision(dataType).build();
-}
-} // namespace
-
-// FIXME: make this thread-safe by reusing the benchmark cache in Conv_v7.cpp
-namespace {
-// we currently set the maximum number of input dimensions to 5
-// this can be increased, if necessary
-constexpr uint8_t max_num_input_dim = 5;
-struct LinearParams {
-  c10::DeviceIndex device_id;
-  cudnnDataType_t dataType;
-  cudnnDataType_t inputDataType;
-  cudnnDataType_t outputDataType;
-  int input_size[max_num_input_dim];
-  uint8_t input_dim;
-  at::MemoryFormat memory_format;
-  int64_t weight_size[2];
-  bool deterministic;
-  bool allow_tf32;
-};
-struct CacheKey {
-  LinearParams params;
-  uint8_t input_alignment;
-  uint8_t weight_alignment;
-  uint8_t output_alignment;
-  // default to -1 when no bias
-  int8_t bias_alignment;
-  bool kReluFused;
-};
-void setLinearParams(LinearParams *params, const at::Tensor &input,
-                     const at::Tensor &weight, bool deterministic,
-                     bool allow_tf32,
-                     cudnnDataType_t inputDataType = CUDNN_DATA_INT8,
-                     cudnnDataType_t floatGEMMComputeType = CUDNN_DATA_FLOAT) {
-  // operator datatype needs to be int32 for int8 matmul, but we can
-  // set the datatype for output tensor to int32 or fp32
-  memset(params, 0, sizeof(LinearParams));
-  params->device_id = at::cuda::current_device();
-  params->inputDataType = inputDataType;
-  params->dataType =
-      inputDataType == CUDNN_DATA_INT8 ? CUDNN_DATA_INT32 : floatGEMMComputeType;
-  params->input_dim = input.dim();
-  params->memory_format = input.suggest_memory_format();
-  for (int i = 0; i < params->input_dim; ++i) {
-    params->input_size[i] = input.sizes()[i];
-  }
-  for (int i = 0; i < 2; ++i) {
-    params->weight_size[i] = weight.sizes()[i];
-  }
-  params->deterministic = deterministic;
-  params->allow_tf32 = allow_tf32;
-}
-std::unordered_map<CacheKey, cudnn_frontend::ExecutionPlan,
-                   at::native::ParamsHash<CacheKey>,
-                   at::native::ParamsEqual<CacheKey>>
-    execution_plan_cache;
-} // namespace
-
-namespace sfast {
-namespace operators {
-// TODO: we can use cudnn_frontend::ExecutionPlanCache when it supports caching
-// multiple operators
-// reference:
-// https://github.com/NVIDIA/cudnn-frontend/blob/main/samples/conv_sample.cpp#L293
-// static cudnn_frontend::ExecutionPlanCache plan_cache("sample_cache");
-
-// currently we only support int8 symmetric (zero_point = 0 for inputs and
-// output) quantized linear op We implement relu(act_int8 * transpose(w_int8) +
-// [bias_fp32/(act_scale * w_scale] ) * ( act_scale * w_scale / out_scale )
-// which requires 5 cudnn ops (1 matmul, 2 multiplication, 1 add, and 1 relu
-// ops) matmul op: linear_op Multiplication ops: rhs_mult_op, requant_op
-// Addition op: add_op
-// Relu op: relu_op
-template <bool kReluFused>
-void PackedLinearWeightCudnn::apply_impl_helper(const at::Tensor& quantized_output, const at::Tensor& input, double output_scale) {
-  if (quantized_output.numel() == 0) {
-    return;
-  }
-  auto act_scale = input.q_scale();
-  auto weight_scale = orig_weight.q_scale();
-  auto requantize_multiplier = act_scale * weight_scale / output_scale;
-  at::Tensor requantize_multiplier_tensor = cudnn_utils::getRequantMultiplierTensor(requantize_multiplier, quantized_output.dim());
-  c10::optional<at::Tensor> bias_multiplier_tensor;
-  c10::optional<at::Tensor> broadcasted_bias;
-  if (bias_.has_value()) {
-    // the input bias is a 1-D tensor whose size is the same as the size of the last dimension of quantized_output
-    // we need to add trailing dimensions in order to properly broadcast bias, otherwise broadcast_to will fail.
-    // the number of trailling dimensions is quantized_output.dim() - 2. We also prepend a leading dimension for clarity
-    std::vector<int64_t> new_size(quantized_output.dim(), 1);
-    new_size.back() = bias_.value().size(0);
-    broadcasted_bias = bias_.value().clone().reshape(new_size);
-    broadcasted_bias.value() = broadcasted_bias.value().broadcast_to(quantized_output.sizes()).contiguous();
-    bias_multiplier_tensor = at::empty(quantized_output.sizes(), at::device(at::kCUDA).dtype(at::kFloat));
-    auto bias_multiplier = 1.0 / (act_scale * weight_scale);
-    bias_multiplier_tensor.value().fill_(bias_multiplier);
-  }
-
-  cudnnHandle_t handle = at::native::getCudnnHandle();
-  CacheKey key;
-  // memset is needed here because there is implicit packing added for CacheKey, and this can result in uninitialized padded values that are
-  // used for hashing (see how at::native::ParamsHash is defined). without memset, we can potentially come across a situation where two
-  // CacheKey objects have the same user defined parameters, but
-  // different padded values, resulting in different hash outputs.
-  memset(&key, 0, sizeof(key));
-  bool deterministic{false};
-  bool allow_tf32{true};
-  setLinearParams(&key.params, input, orig_weight, deterministic, allow_tf32);
-
-  key.input_alignment = cudnn_utils::getAlignment(input);
-  key.output_alignment = cudnn_utils::getAlignment(quantized_output);
-  key.weight_alignment = cudnn_utils::getAlignment(orig_weight);
-  if (bias_.has_value()) {
-    key.bias_alignment = cudnn_utils::getAlignment(broadcasted_bias.value());
-  } else {
-    key.bias_alignment = -1;
-  }
-  key.kReluFused = kReluFused;
-  // the matmul operation is input * transpose(weight), so we will work with the transposed weight
-  auto weight_transposed = transpose(orig_weight, 0, 1);
-  // cudnn expects tensors to be at least 3D. weight_transposed is currently 2D. we will create a 3D view
-  // by prepending a leading dummy dimension (cudnn expects leading dimensions to be the dummy dimensions)
-  std::vector<int64_t> new_sizes(3, 1);
-  new_sizes.back() = weight_transposed.size(1);
-  new_sizes[1] = weight_transposed.size(0);
-  weight_transposed = weight_transposed.view(new_sizes);
-
-  auto run = [&](const cudnn_frontend::ExecutionPlan& plan_desc) {
-    auto workspace_size = plan_desc.getWorkspaceSize();
-    auto workspace_ptr = c10::cuda::CUDACachingAllocator::get()->allocate(workspace_size);
-    at::SmallVector<void *, 8> data_ptrs;
-    at::SmallVector<int64_t, 8> uids;
-    data_ptrs = {input.data_ptr<int8_t>(), weight_transposed.data_ptr<int8_t>(),
-                 requantize_multiplier_tensor.data_ptr(), quantized_output.data_ptr<int8_t>()};
-    uids = {'x', 'w', 's', 'r'};
-    if (bias_.has_value()) {
-      data_ptrs.insert(data_ptrs.end(), {broadcasted_bias.value().data_ptr(), bias_multiplier_tensor.value().data_ptr(),
-                                         broadcasted_bias.value().data_ptr(), broadcasted_bias.value().data_ptr()});
-      uids.insert(uids.end(), {'b', 'c', 'd', 'n'});
-    }
-    auto variantPack = cudnn_frontend::VariantPackBuilder()
-      .setWorkspacePointer(workspace_size ? workspace_ptr.get() : nullptr)
-      .setDataPointers(uids.size(), data_ptrs.data())
-      .setUids(uids.size(), uids.data())
-      .build();
-    auto variant_pack_desc = variantPack.get_raw_desc();
-    AT_CUDNN_CHECK(cudnnBackendExecute(handle, plan_desc.get_raw_desc(), variant_pack_desc));
-  };
-
-  auto search = execution_plan_cache.find(key);
-  if (search != execution_plan_cache.end()) {
-    cudnn_frontend::ExecutionPlan plan_desc = search->second;
-    run(plan_desc);
-    return;
-  }
-
-  // linear_op computes act_int8 * tranpose(w_int8) (matrix multiplication)
-  // where act_int8 and w_int8 are the input and weight variables, resp.
-  // output is a fp32 tensor
-  auto linear_op = cudnn_frontend::OperationBuilder(CUDNN_BACKEND_OPERATION_MATMUL_DESCRIPTOR)
-      .setaMatDesc(cudnn_utils::getTensorDescriptor(input.sizes(), input.strides(), CUDNN_DATA_INT8, 'x', key.input_alignment))
-      .setbMatDesc(cudnn_utils::getTensorDescriptor(weight_transposed.sizes(), weight_transposed.strides(), CUDNN_DATA_INT8, 'w', key.weight_alignment))
-      // for virtual tensors, the alignment is not used, so we can just put an arbitrary value here, e.g., key.output_alignment
-      .setcMatDesc(cudnn_utils::getTensorDescriptor(quantized_output.sizes(), quantized_output.strides(), CUDNN_DATA_FLOAT, 'y', key.output_alignment, true))
-      .setmatmulDesc(getLinearDescriptor(key.params.dataType))
-      .build();
-  // std::cout << "operator:" << linear_op.describe() << std::endl;
-
-  c10::optional<cudnn_frontend::Operation> bias_mult_op;
-  c10::optional<cudnn_frontend::Operation> sum_linear_bias_op;
-  if (bias_.has_value()) {
-    // we can't directly assign bias_mult_op becauase operator= is deleted for cudnn_frontend::Operation;
-    // alternatively, I think we can use std::unique_ptr and dynamically allocate these builder ops
-    // but here, we chose to do it statically. c10::optional<T>::emplace() enables this approach
-
-    // bias_mult_op computes bias_fp32 / (act_scale * w_scale) or bias_fp32 * (1 / (act_scale * w_scale))
-    // where bias_multiplier = (1 / (act_scale * w_scale))
-    // output is a fp32 tensor
-    // we use inplace operation here where the output is assigned to the input
-    bias_mult_op.emplace(cudnn_frontend::OperationBuilder(CUDNN_BACKEND_OPERATION_POINTWISE_DESCRIPTOR)
-      .setxDesc(cudnn_utils::getTensorDescriptor(broadcasted_bias.value(), 'b', cudnn_utils::getAlignment(broadcasted_bias.value())))
-      .setbDesc(cudnn_utils::getTensorDescriptor(bias_multiplier_tensor.value(), 'c', cudnn_utils::getAlignment(bias_multiplier_tensor.value())))
-      // TODO: I think we should be able to make this a virtual tensor, but we would need cudnn to support
-      // setbdesc(ManagedOpaqueDescriptor const &raw_tensor) first
-      .setyDesc(cudnn_utils::getTensorDescriptor(broadcasted_bias.value(), 'd', cudnn_utils::getAlignment(broadcasted_bias.value())))
-      .setpwDesc(cudnn_utils::getPointWiseMulDescriptor(getCudnnDataType(bias_multiplier_tensor.value())))
-      .build());
-
-    // computes (act_int8 * w_int8 + [bias_fp32/(act_scale * w_scale)])
-    // where the 1st and 2nd summands is output of linear op and broadcasted_bias, resp.
-    // output is a fp32 tensor
-    // we use inplace operation here where the output is assigned to the input
-    sum_linear_bias_op.emplace(cudnn_frontend::OperationBuilder(CUDNN_BACKEND_OPERATION_POINTWISE_DESCRIPTOR)
-      .setxDesc(linear_op.getOutputTensor())
-      // TODO: An additional entry for broadcasted_bias in the uid-data_ptr pairing
-      // appears to be needed in the current version of cudnn (8.4.0). Without it, some
-      // test cases are failing. NVIDIA is currently investigating this issue.
-      // When this issue is fixed, we can change 'n' back to 'd' and remove the additional entry in uid and data_ptrs in variant pack above
-      .setbDesc(cudnn_utils::getTensorDescriptor(broadcasted_bias.value(), 'n', cudnn_utils::getAlignment(broadcasted_bias.value())))
-      .setyDesc(cudnn_utils::getTensorDescriptor(quantized_output.sizes(), quantized_output.strides(), CUDNN_DATA_FLOAT, 'e', key.output_alignment, true))
-      .setpwDesc(cudnn_utils::getPointWiseAddDescriptor(getCudnnDataType(broadcasted_bias.value())))
-      .build());
-  }
-
-  // relu_op computes relu(act_int8 * w_int8 + [bias_fp32/(act_scale * w_scale)]
-  // or relu(act_int8 * w_int8) if bias is not present.
-  // output is a fp32 tensor
-  c10::optional<cudnn_frontend::Operation> relu_op;
-  std::shared_ptr<cudnn_frontend::OpaqueBackendPointer> tensor2requant_ptr = bias_.has_value() ? sum_linear_bias_op.value().getOutputTensor() : linear_op.getOutputTensor();
-  if (kReluFused) {
-    // we use inplace operation here where the output is assigned to the input
-    relu_op.emplace(cudnn_frontend::OperationBuilder(CUDNN_BACKEND_OPERATION_POINTWISE_DESCRIPTOR)
-      .setxDesc(tensor2requant_ptr)
-      // for virtual tensors, the alignment is not used, so we can just put an arbitrary value here, e.g., key.output_alignment
-      .setyDesc(cudnn_utils::getTensorDescriptor(quantized_output.sizes(), quantized_output.strides(), CUDNN_DATA_FLOAT, 'f', key.output_alignment, true))
-      .setpwDesc(cudnn_utils::getPointWiseReluDescriptor(CUDNN_DATA_FLOAT))
-      .build());
-  }
-
-  // requant_op computes relu(act_int8 * w_int8 + [bias_fp32/(act_scale * w_scale)]) / (out_scale / (act_scale * w_scale))
-  // or relu(act_int8 * w_int8) / (out_scale / (act_scale * w_scale))) if bias is not present.
-  // output is a fp32 tensor
-  auto requant_op = cudnn_frontend::OperationBuilder(CUDNN_BACKEND_OPERATION_POINTWISE_DESCRIPTOR)
-    .setxDesc(kReluFused ? relu_op.value().getOutputTensor() : tensor2requant_ptr)
-    .setbDesc(cudnn_utils::getTensorDescriptor(requantize_multiplier_tensor, 's', cudnn_utils::getAlignment(requantize_multiplier_tensor)))
-    .setyDesc(cudnn_utils::getTensorDescriptor(quantized_output.sizes(), quantized_output.strides(), CUDNN_DATA_INT8, 'r', key.output_alignment))
-    .setpwDesc(cudnn_utils::getPointWiseMulDescriptor(getCudnnDataType(requantize_multiplier_tensor)))
-    .build();
-  // // std::cout << "operator:" << requant_op.describe() << std::endl;
-
-  std::vector<cudnn_frontend::Operation const *> ops{&linear_op};
-  if (bias_.has_value()) {
-    ops.emplace_back(&(bias_mult_op.value()));
-    ops.emplace_back(&(sum_linear_bias_op.value()));
-  }
-  if (kReluFused) {
-    ops.emplace_back(&(relu_op.value()));
-  }
-  ops.emplace_back(&requant_op);
-
-  auto opGraph = cudnn_frontend::OperationGraphBuilder()
-      .setHandle(handle)
-      .setOperationGraph(ops.size(), ops.data())
-      .build();
-  // std::cout << "opGraph: " << opGraph.describe() << std::endl;
-
-  auto heuristics = cudnn_frontend::EngineHeuristicsBuilder()
-      .setOperationGraph(opGraph)
-      .setHeurMode(CUDNN_HEUR_MODE_INSTANT)
-      .build();
-  auto fallback = cudnn_frontend::EngineFallbackListBuilder()
-                    .setOperationGraph(opGraph)
-                    .setOperation(CUDNN_BACKEND_OPERATION_MATMUL_DESCRIPTOR)
-                    .build();
-
-  auto& engine_configs = heuristics.getEngineConfig(heuristics.getEngineConfigCount());
-  auto& fallback_list = fallback.getFallbackList();
-
-  cudnn_frontend::EngineConfigList filtered_configs;
-  cudnn_utils::filterEngineConfigs(engine_configs, filtered_configs, deterministic, allow_tf32, at::kChar);
-  cudnn_utils::filterEngineConfigs(fallback_list, filtered_configs, deterministic, allow_tf32, at::kChar);
-
-  for (auto &cfg : filtered_configs) {
-    try {
-      auto plan = cudnn_frontend::ExecutionPlanBuilder()
-        .setHandle(handle)
-        .setEngineConfig(cfg)
-        .build();
-      run(plan);
-      execution_plan_cache.emplace(key, plan);
-      return;
-    } catch (cudnn_frontend::cudnnException &e) {std::cout << "cudnn error:" << e.what() << std::endl;} catch(c10::CuDNNError &e) { std::cout << "other error" << e.what() << std::endl;}
-  }
-
-  TORCH_CHECK(false, "Unable to find an engine to execute this computation Quantized Linear Cudnn");
-}
-
-template <bool kReluFused>
-void PackedLinearWeightCudnn::apply_impl_helper_dynamic(const at::Tensor& output, const at::Tensor& input) {
-  if (output.numel() == 0) {
-    return;
-  }
-
-  auto floatScalarType = input.scalar_type();
-  auto floatGEMMComputeScalarType = floatScalarType == at::kBFloat16 ? at::kFloat : floatScalarType;
-  auto floatDataType = getCudnnDataType(input);
-  auto floatGEMMComputeType = floatDataType == CUDNN_DATA_BFLOAT16 ? CUDNN_DATA_FLOAT : floatDataType;
-
-  auto output_transposed = transpose(output, -2, -1);
-  auto weight_scale = orig_weight.q_scale();
-  auto requantize_multiplier = weight_scale;
-  at::Tensor requantize_multiplier_tensor = cudnn_utils::getRequantMultiplierTensor(requantize_multiplier, output_transposed.dim());
-  c10::optional<at::Tensor> bias_multiplier_tensor;
-  c10::optional<at::Tensor> broadcasted_bias;
-  if (bias_.has_value()) {
-    // the input bias is a 1-D tensor whose size is the same as the size of the last dimension of quantized_output
-    // we need to add trailing dimensions in order to properly broadcast bias, otherwise broadcast_to will fail.
-    // the number of trailling dimensions is quantized_output.dim() - 2. We also prepend a leading dimension for clarity
-    std::vector<int64_t> new_size(output_transposed.dim(), 1);
-    new_size.back() = bias_.value().size(0);
-    broadcasted_bias = bias_.value().clone().reshape(new_size);
-    broadcasted_bias.value() = broadcasted_bias.value().broadcast_to(output_transposed.sizes()).contiguous();
-    bias_multiplier_tensor = at::empty(output_transposed.sizes(), at::device(at::kCUDA).dtype(at::kFloat));
-    auto bias_multiplier = 1.0 / weight_scale;
-    bias_multiplier_tensor.value().fill_(bias_multiplier);
-  }
-
-  cudnnHandle_t handle = at::native::getCudnnHandle();
-  CacheKey key;
-  // memset is needed here because there is implicit packing added for CacheKey, and this can result in uninitialized padded values that are
-  // used for hashing (see how at::native::ParamsHash is defined). without memset, we can potentially come across a situation where two
-  // CacheKey objects have the same user defined parameters, but
-  // different padded values, resulting in different hash outputs.
-  memset(&key, 0, sizeof(key));
-  bool deterministic{false};
-  bool allow_tf32{true};
-  setLinearParams(&key.params, input, orig_weight, deterministic, allow_tf32, floatDataType, floatGEMMComputeType);
-
-  key.input_alignment = cudnn_utils::getAlignment(input);
-  key.output_alignment = cudnn_utils::getAlignment(output);
-  key.weight_alignment = cudnn_utils::getAlignment(orig_weight);
-  if (bias_.has_value()) {
-    key.bias_alignment = cudnn_utils::getAlignment(broadcasted_bias.value());
-  } else {
-    key.bias_alignment = -1;
-  }
-  key.kReluFused = kReluFused;
-  // the matmul operation is input * transpose(weight), so we will work with the transposed weight
-  auto input_transposed = transpose(input, -2, -1);
-  // cudnn expects tensors to be at least 3D. weight_transposed is currently 2D. we will create a 3D view
-  // by prepending a leading dummy dimension (cudnn expects leading dimensions to be the dummy dimensions)
-  std::vector<int64_t> new_sizes(3, 1);
-  new_sizes.back() = orig_weight.size(1);
-  new_sizes[1] = orig_weight.size(0);
-  auto weight = orig_weight.view(new_sizes);
-
-  auto run = [&](const cudnn_frontend::ExecutionPlan& plan_desc) {
-    auto workspace_size = plan_desc.getWorkspaceSize();
-    auto workspace_ptr = c10::cuda::CUDACachingAllocator::get()->allocate(workspace_size);
-    at::SmallVector<void *, 8> data_ptrs;
-    at::SmallVector<int64_t, 8> uids;
-    data_ptrs = {input_transposed.data_ptr(), weight.data_ptr<int8_t>(),
-                 requantize_multiplier_tensor.data_ptr(), output_transposed.data_ptr()};
-    uids = {'x', 'w', 's', 'r'};
-    if (bias_.has_value()) {
-      data_ptrs.insert(data_ptrs.end(), {broadcasted_bias.value().data_ptr(), bias_multiplier_tensor.value().data_ptr(),
-                                         broadcasted_bias.value().data_ptr(), broadcasted_bias.value().data_ptr()});
-      uids.insert(uids.end(), {'b', 'c', 'd', 'n'});
-    }
-    auto variantPack = cudnn_frontend::VariantPackBuilder()
-      .setWorkspacePointer(workspace_size ? workspace_ptr.get() : nullptr)
-      .setDataPointers(uids.size(), data_ptrs.data())
-      .setUids(uids.size(), uids.data())
-      .build();
-    auto variant_pack_desc = variantPack.get_raw_desc();
-    AT_CUDNN_CHECK(cudnnBackendExecute(handle, plan_desc.get_raw_desc(), variant_pack_desc));
-  };
-
-  auto search = execution_plan_cache.find(key);
-  if (search != execution_plan_cache.end()) {
-    cudnn_frontend::ExecutionPlan plan_desc = search->second;
-    run(plan_desc);
-    return;
-  }
-
-  auto weight_to_float_op = cudnn_frontend::OperationBuilder(CUDNN_BACKEND_OPERATION_POINTWISE_DESCRIPTOR)
-    .setxDesc(cudnn_utils::getTensorDescriptor(weight.sizes(), weight.strides(), CUDNN_DATA_INT8, 'w', key.weight_alignment))
-    .setyDesc(cudnn_utils::getTensorDescriptor(weight.sizes(), weight.strides(), floatDataType, 'z', key.weight_alignment, true))
-    .setpwDesc(cudnn_utils::getPointWiseIdentityDescriptor(floatDataType))
-    .build();
-
-  // linear_op computes act_int8 * tranpose(w_int8) (matrix multiplication)
-  // where act_int8 and w_int8 are the input and weight variables, resp.
-  // output is a fp32 tensor
-  auto linear_op = cudnn_frontend::OperationBuilder(CUDNN_BACKEND_OPERATION_MATMUL_DESCRIPTOR)
-      .setaMatDesc(weight_to_float_op.getOutputTensor())
-      .setbMatDesc(cudnn_utils::getTensorDescriptor(input_transposed.sizes(), input_transposed.strides(), floatDataType, 'x', key.input_alignment))
-      // for virtual tensors, the alignment is not used, so we can just put an arbitrary value here, e.g., key.output_alignment
-      .setcMatDesc(cudnn_utils::getTensorDescriptor(output_transposed.sizes(), output_transposed.strides(), floatDataType, 'y', key.output_alignment, true))
-      .setmatmulDesc(getLinearDescriptor(key.params.dataType))
-      .build();
-  // std::cout << "operator:" << linear_op.describe() << std::endl;
-
-  c10::optional<cudnn_frontend::Operation> bias_mult_op;
-  c10::optional<cudnn_frontend::Operation> sum_linear_bias_op;
-  if (bias_.has_value()) {
-    // we can't directly assign bias_mult_op becauase operator= is deleted for cudnn_frontend::Operation;
-    // alternatively, I think we can use std::unique_ptr and dynamically allocate these builder ops
-    // but here, we chose to do it statically. c10::optional<T>::emplace() enables this approach
-
-    // bias_mult_op computes bias_fp32 / (act_scale * w_scale) or bias_fp32 * (1 / (act_scale * w_scale))
-    // where bias_multiplier = (1 / (act_scale * w_scale))
-    // output is a fp32 tensor
-    // we use inplace operation here where the output is assigned to the input
-    bias_mult_op.emplace(cudnn_frontend::OperationBuilder(CUDNN_BACKEND_OPERATION_POINTWISE_DESCRIPTOR)
-      .setxDesc(cudnn_utils::getTensorDescriptor(broadcasted_bias.value(), 'b', cudnn_utils::getAlignment(broadcasted_bias.value())))
-      .setbDesc(cudnn_utils::getTensorDescriptor(bias_multiplier_tensor.value(), 'c', cudnn_utils::getAlignment(bias_multiplier_tensor.value())))
-      // TODO: I think we should be able to make this a virtual tensor, but we would need cudnn to support
-      // setbdesc(ManagedOpaqueDescriptor const &raw_tensor) first
-      .setyDesc(cudnn_utils::getTensorDescriptor(broadcasted_bias.value(), 'd', cudnn_utils::getAlignment(broadcasted_bias.value())))
-      .setpwDesc(cudnn_utils::getPointWiseMulDescriptor(getCudnnDataType(bias_multiplier_tensor.value())))
-      .build());
-
-    // computes (act_int8 * w_int8 + [bias_fp32/(act_scale * w_scale)])
-    // where the 1st and 2nd summands is output of linear op and broadcasted_bias, resp.
-    // output is a fp32 tensor
-    // we use inplace operation here where the output is assigned to the input
-    sum_linear_bias_op.emplace(cudnn_frontend::OperationBuilder(CUDNN_BACKEND_OPERATION_POINTWISE_DESCRIPTOR)
-      .setxDesc(linear_op.getOutputTensor())
-      // TODO: An additional entry for broadcasted_bias in the uid-data_ptr pairing
-      // appears to be needed in the current version of cudnn (8.4.0). Without it, some
-      // test cases are failing. NVIDIA is currently investigating this issue.
-      // When this issue is fixed, we can change 'n' back to 'd' and remove the additional entry in uid and data_ptrs in variant pack above
-      .setbDesc(cudnn_utils::getTensorDescriptor(broadcasted_bias.value(), 'n', cudnn_utils::getAlignment(broadcasted_bias.value())))
-      .setyDesc(cudnn_utils::getTensorDescriptor(output.sizes(), output.strides(), CUDNN_DATA_FLOAT, 'e', key.output_alignment, true))
-      .setpwDesc(cudnn_utils::getPointWiseAddDescriptor(getCudnnDataType(broadcasted_bias.value())))
-      .build());
-  }
-
-  // relu_op computes relu(act_int8 * w_int8 + [bias_fp32/(act_scale * w_scale)]
-  // or relu(act_int8 * w_int8) if bias is not present.
-  // output is a fp32 tensor
-  c10::optional<cudnn_frontend::Operation> relu_op;
-  std::shared_ptr<cudnn_frontend::OpaqueBackendPointer> tensor2requant_ptr = bias_.has_value() ? sum_linear_bias_op.value().getOutputTensor() : linear_op.getOutputTensor();
-  if (kReluFused) {
-    // we use inplace operation here where the output is assigned to the input
-    relu_op.emplace(cudnn_frontend::OperationBuilder(CUDNN_BACKEND_OPERATION_POINTWISE_DESCRIPTOR)
-      .setxDesc(tensor2requant_ptr)
-      // for virtual tensors, the alignment is not used, so we can just put an arbitrary value here, e.g., key.output_alignment
-      .setyDesc(cudnn_utils::getTensorDescriptor(output_transposed.sizes(), output_transposed.strides(), floatDataType, 'f', key.output_alignment, true))
-      .setpwDesc(cudnn_utils::getPointWiseReluDescriptor(CUDNN_DATA_FLOAT))
-      .build());
-  }
-
-  // requant_op computes relu(act_int8 * w_int8 + [bias_fp32/(act_scale * w_scale)]) / (out_scale / (act_scale * w_scale))
-  // or relu(act_int8 * w_int8) / (out_scale / (act_scale * w_scale))) if bias is not present.
-  // output is a fp32 tensor
-  auto requant_op = cudnn_frontend::OperationBuilder(CUDNN_BACKEND_OPERATION_POINTWISE_DESCRIPTOR)
-    .setxDesc(kReluFused ? relu_op.value().getOutputTensor() : tensor2requant_ptr)
-    .setbDesc(cudnn_utils::getTensorDescriptor(requantize_multiplier_tensor, 's', cudnn_utils::getAlignment(requantize_multiplier_tensor)))
-    .setyDesc(cudnn_utils::getTensorDescriptor(output_transposed.sizes(), output_transposed.strides(), floatDataType, 'r', key.output_alignment))
-    .setpwDesc(cudnn_utils::getPointWiseMulDescriptor(getCudnnDataType(requantize_multiplier_tensor)))
-    .build();
-  // // std::cout << "operator:" << requant_op.describe() << std::endl;
-
-  std::vector<cudnn_frontend::Operation const *> ops{&weight_to_float_op, &linear_op};
-  if (bias_.has_value()) {
-    ops.emplace_back(&(bias_mult_op.value()));
-    ops.emplace_back(&(sum_linear_bias_op.value()));
-  }
-  if (kReluFused) {
-    ops.emplace_back(&(relu_op.value()));
-  }
-  ops.emplace_back(&requant_op);
-
-  auto opGraph = cudnn_frontend::OperationGraphBuilder()
-      .setHandle(handle)
-      .setOperationGraph(ops.size(), ops.data())
-      .build();
-  // std::cout << "opGraph: " << opGraph.describe() << std::endl;
-
-  auto heuristics = cudnn_frontend::EngineHeuristicsBuilder()
-      .setOperationGraph(opGraph)
-      .setHeurMode(CUDNN_HEUR_MODE_INSTANT)
-      .build();
-  auto fallback = cudnn_frontend::EngineFallbackListBuilder()
-                    .setOperationGraph(opGraph)
-                    .setOperation(CUDNN_BACKEND_OPERATION_MATMUL_DESCRIPTOR)
-                    .build();
-
-  auto& engine_configs = heuristics.getEngineConfig(heuristics.getEngineConfigCount());
-  auto& fallback_list = fallback.getFallbackList();
-
-  cudnn_frontend::EngineConfigList filtered_configs;
-  cudnn_utils::filterEngineConfigs(engine_configs, filtered_configs, deterministic, allow_tf32, floatGEMMComputeScalarType);
-  cudnn_utils::filterEngineConfigs(fallback_list, filtered_configs, deterministic, allow_tf32, floatGEMMComputeScalarType);
-
-  for (auto &cfg : filtered_configs) {
-    try {
-      auto plan = cudnn_frontend::ExecutionPlanBuilder()
-        .setHandle(handle)
-        .setEngineConfig(cfg)
-        .build();
-      run(plan);
-      execution_plan_cache.emplace(key, plan);
-      return;
-    } catch (cudnn_frontend::cudnnException &e) {std::cout << "cudnn error:" << e.what() << std::endl;} catch(c10::CuDNNError &e) { std::cout << "other error" << e.what() << std::endl;}
-  }
-
-  TORCH_CHECK(false, "Unable to find an engine to execute this computation Quantized Linear Cudnn");
-}
-
-// output Tensor will be a clampped int8 Tensor
-// both act and weight will be int8 Tensor
-// Numerics are the same as conv (see aten/src/ATen/native/quantized/Conv.cpp):
-template <bool kReluFused>
-at::Tensor PackedLinearWeightCudnn::apply_impl(const at::Tensor &act,
-                                               double output_scale,
-                                               int64_t output_zero_point) {
-  TORCH_CHECK(act.device().is_cuda(), "act must be a CUDA tensor");
-  TORCH_CHECK(orig_weight.device().is_cuda(), "weight must be a CUDA tensor");
-  TORCH_CHECK(!bias_.has_value() || bias_.value().device().is_cuda(),
-              "bias must be a CUDA tensor");
-
-  std::vector<int64_t> original_output_shape{act.sizes().vec()}; // 2D
-  original_output_shape.back() = orig_weight.size(0); // output channels
-  // cudnn expects tensors to be at least 3D. we will prepend a dummy dimension
-  // for quantized_output
-  std::vector<int64_t> output_shape(3, 1);
-  output_shape[1] = original_output_shape[0];
-  output_shape[2] = original_output_shape[1];
-  // cudnn expects tensors to be at least 3D. act is currently 2D. we will
-  // create a 3D view
-  std::vector<int64_t> new_sizes(3, 1);
-  // cudnn expects leading dimensions to be the dummy dimensions
-  new_sizes.back() = act.sizes().back();
-  new_sizes[1] = act.size(0);
-  if (act.is_floating_point()) {
-    std::swap(output_shape[1], output_shape[2]);
-    auto output = act.new_empty(output_shape);
-    output = transpose(output, -2, -1);
-    apply_impl_helper_dynamic<kReluFused>(output, act.view(new_sizes));
-    return output.view(original_output_shape);
-  } else {
-    auto quantized_output = at::_empty_affine_quantized(
-        output_shape, at::device(at::kCUDA).dtype(at::ScalarType::QInt8),
-        output_scale, output_zero_point);
-    apply_impl_helper<kReluFused>(quantized_output, act.view(new_sizes),
-                                  output_scale);
-    return quantized_output.view(original_output_shape);
-  }
-}
-
-at::Tensor PackedLinearWeightCudnn::apply(at::Tensor input, double output_scale,
-                                          int64_t output_zero_point) {
-  return apply_impl<false>(input, output_scale, output_zero_point);
-}
-
-at::Tensor PackedLinearWeightCudnn::apply_relu(at::Tensor input,
-                                               double output_scale,
-                                               int64_t output_zero_point) {
-  return apply_impl<true>(input, output_scale, output_zero_point);
-}
-
-at::Tensor PackedLinearWeightCudnn::apply_dynamic(at::Tensor input,
-                                                  bool reduce_range) {
-  if (reduce_range) {
-    TORCH_WARN(
-        "Currently CUDNN QLinear ignores reduce_range when it its set to true");
-  }
-  return apply_impl<false>(input);
-}
-
-at::Tensor PackedLinearWeightCudnn::apply_dynamic_relu(at::Tensor input,
-                                                       bool reduce_range) {
-  if (reduce_range) {
-    TORCH_WARN(
-        "Currently CUDNN QLinear ignores reduce_range when it its set to true");
-  }
-  return apply_impl<true>(input);
-}
-
-std::tuple<at::Tensor, c10::optional<at::Tensor>>
-PackedLinearWeightCudnn::unpack() {
-  return std::tuple<at::Tensor, c10::optional<at::Tensor>>{orig_weight, bias_};
-}
-
-c10::intrusive_ptr<LinearPackedParamsBase>
-PackedLinearWeightCudnn::prepack(at::Tensor weight,
-                                 c10::optional<at::Tensor> bias) {
-  TORCH_CHECK(weight.qscheme() == c10::kPerTensorAffine,
-              "Unsupported qscheme: ", toString(weight.qscheme()));
-  const int output_channels = weight.size(0);
-  const auto qtype = weight.qscheme();
-  if (bias.has_value()) {
-    TORCH_CHECK(bias.value().dim() == 1, "bias should be a vector (1D Tensor)");
-    TORCH_CHECK(bias.value().size(0) == output_channels,
-                "bias should have K elements: " +
-                    std::to_string(output_channels));
-  }
-
-  auto ret_ptr =
-      c10::make_intrusive<PackedLinearWeightCudnn>(weight, bias, qtype);
-  return ret_ptr;
-}
-
-c10::intrusive_ptr<LinearPackedParamsBase> PackedLinearWeightCudnn::from_native(
-    const c10::intrusive_ptr<LinearPackedParamsBase> &other) {
-  if (dynamic_cast<PackedLinearWeightCudnn *>(other.get()) != nullptr) {
-    return other;
-  }
-  auto unpacked = other->unpack();
-  auto orig_weight = std::get<0>(unpacked);
-  auto orig_bias = std::get<1>(unpacked);
-  return PackedLinearWeightCudnn::prepack(orig_weight, orig_bias);
-}
-
-namespace {
-
-template <bool kReluFused> class QLinearInt8 final {
-public:
-  static at::Tensor
-  run(at::Tensor act,
-      const c10::intrusive_ptr<LinearPackedParamsBase> &packed_weight,
-      double output_scale, int64_t output_zero_point) {
-    // TODO: check all zero_points are zero/all tensors are symmetrically
-    // quantized
-    auto packed_weight_cudnn =
-        PackedLinearWeightCudnn::from_native(packed_weight);
-    if (kReluFused) {
-      return packed_weight_cudnn->apply_relu(act, output_scale,
-                                             output_zero_point);
-    } else {
-      return packed_weight_cudnn->apply(act, output_scale, output_zero_point);
-    }
-  }
-
-  static at::Tensor
-  run_dynamic(at::Tensor act,
-              const c10::intrusive_ptr<LinearPackedParamsBase> &packed_weight,
-              bool reduce_range = false) {
-    auto packed_weight_cudnn =
-        PackedLinearWeightCudnn::from_native(packed_weight);
-    if (kReluFused) {
-      return packed_weight_cudnn->apply_dynamic_relu(act);
-    } else {
-      return packed_weight_cudnn->apply_dynamic(act);
-    }
-  }
-};
-
-TORCH_LIBRARY_IMPL(quantized, QuantizedCUDA, m) {
-  m.impl(TORCH_SELECTIVE_NAME("quantized::linear_dynamic"),
-         QLinearInt8<false>::run_dynamic);
-  m.impl(TORCH_SELECTIVE_NAME("quantized::linear_relu_dynamic"),
-         QLinearInt8<true>::run_dynamic);
-}
-
-TORCH_LIBRARY_IMPL(quantized, CUDA, m) {
-  m.impl(TORCH_SELECTIVE_NAME("quantized::linear_dynamic"),
-         QLinearInt8<false>::run_dynamic);
-  m.impl(TORCH_SELECTIVE_NAME("quantized::linear_relu_dynamic"),
-         QLinearInt8<true>::run_dynamic);
-}
-
-} // namespace
-
-void initCUDNNQLinearBindings(torch::Library &m) {
-  m.def("cudnn_qlinear",
-        torch::dispatch(c10::DispatchKey::CompositeImplicitAutograd,
-                        QLinearInt8<false>::run));
-  m.def("cudnn_qlinear_relu",
-        torch::dispatch(c10::DispatchKey::CompositeImplicitAutograd,
-                        QLinearInt8<true>::run));
-  m.def("cudnn_qlinear_dynamic",
-        torch::dispatch(c10::DispatchKey::CompositeImplicitAutograd,
-                        QLinearInt8<false>::run_dynamic));
-  m.def("cudnn_qlinear_relu_dynamic",
-        torch::dispatch(c10::DispatchKey::CompositeImplicitAutograd,
-                        QLinearInt8<true>::run_dynamic));
-}
-
-} // namespace operators
-} // namespace sfast
