+++ sfast/triton/ops/activation.py
@@ -10,3 +10,8 @@ def identity(x):
 @triton.jit
 def silu(x):
     return x * tl.sigmoid(x)
+
+
+@triton.jit
+def relu(x):
+    return tl.max(x, 0)
